% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{changepoint.plot}
\alias{changepoint.plot}
\title{Changepoint outlier detection plot - Only intended for developer use}
\usage{
changepoint.plot(x, show.segments, ...)
}
\arguments{
\item{x}{a list obtained as an output of function \code{\link{KRDetect.outliers.changepoint}} for identification of outliers using changepoint analysis.}

\item{show.segments}{a logical variable specifying if vertical lines representing individual segments are plotted.}

\item{...}{further arguments to be passed to the \code{\link{plot}} function.}
}
\description{
Plot of results obtained using function \code{\link{KRDetect.outliers.changepoint}} for identification of outliers using changepoint analysis.
The function is called by \code{\link{plot.KRDetect}} and is not intended for use by regular users of the package.
}
\details{
This function plots the results obtained using function \code{\link{KRDetect.outliers.changepoint}} identificating outliers using changepoint analysis based method.
The function is exported for developer use only. It does not perform any checks on inputs since it is only convenience function for plotting results obtained using functions implemented in package \pkg{envoutliers}.
}
