% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{controlchart.plot}
\alias{controlchart.plot}
\title{Control chart outliers detection plot - Only intended for developer use}
\usage{
controlchart.plot(x, plot.type = "all", ...)
}
\arguments{
\item{x}{a list obtained as an output of function \code{\link{KRDetect.outliers.controlchart}} for identification of outliers using control charts.}

\item{plot.type}{a type of plot with outliers displayed.

Possible options are
\itemize{
  \item{\code{"all"}} {to show outliers detected using control chart \emph{x}, \emph{R} and \emph{s}}
  \item{\code{"x"}} {to show outliers detected using control chart \emph{x}}
  \item{\code{"R"}} {to show outliers detected using control chart \emph{R}}
  \item{\code{"s"}} {to show outliers detected using control chart \emph{s}}
}}

\item{...}{further arguments to be passed to the \code{\link{plot}} function.}
}
\description{
Plot of results obtained using function \code{\link{KRDetect.outliers.controlchart}} for identification of outliers using control charts.
The function is called by \code{\link{plot.KRDetect}} and is not intended for use by regular users of the package.
}
\details{
This function plots the results obtained using function \code{\link{KRDetect.outliers.controlchart}} identificating outliers using control charts.
The function is exported for developer use only. It does not perform any checks on inputs since it is only convenience function for plotting results obtained using functions implemented in package \pkg{envoutliers}.
}
