% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremevalue.R
\name{extremal.index.gomes}
\alias{extremal.index.gomes}
\title{Extremal index estimation - Only intended for developer use}
\usage{
extremal.index.gomes(x, block.length)
}
\arguments{
\item{x}{a numeric vector of observations.}

\item{block.length}{a numeric value giving the length of blocks.}
}
\value{
A numeric value of an extremal index
}
\description{
Estimation of an extremal index using the block maxima approach suggested by (Gomes, 1993).
The function is called by \code{\link{KRDetect.outliers.EV}} and is not intended for use by regular users of the package.
}
\details{
This function computes the estimate of extremal index suggested by (Gomes, 1993).
The function is exported for developer use only. It does not perform any checks on inputs since it is only convenience function used within \code{\link{KRDetect.outliers.EV}}.
}
\references{
Gomes M (1993). On the estimation of parameter of rare events in environmental time series. In Statistics for the Environment, volume 2 of Water Related Issues, pp. 225-241. Wiley.

Heffernan JE, Stephenson AG (2016). ismev: An Introduction to Statistical Modeling of Extreme Values. R package version 1.41, URL http://CRAN.R-project.org/package=ismev.
}
