% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremevalue.R
\name{extremal.index.intervals}
\alias{extremal.index.intervals}
\title{Extremal index estimation (Ferro and Segers, 2003) - Only intended for developer use}
\usage{
extremal.index.intervals(x, u)
}
\arguments{
\item{x}{a numeric vector of observations.}

\item{u}{a numeric value giving threshold}
}
\value{
a numeric value of an extremal index estimate
}
\description{
Estimation of an extremal index using the Intervals estimator suggested in (Ferro and Segers, 2003).
The function is called by \code{\link{KRDetect.outliers.EV}} and is not intended for use by regular users of the package.
}
\details{
This function computes the estimate of extremal index suggested in (Ferro and Segers, 2003).
The function is exported for developer use only. It does not perform any checks on inputs since it is only convenience function used within \code{\link{KRDetect.outliers.EV}}.
}
\references{
Ferro, CAT, Segers, J (2003). Inference for Cluster of Extreme Values. Journal of Royal Statistical Society, Series B, 65(2), 545-556.
}
