% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlchart.R
\name{get.norm}
\alias{get.norm}
\title{Table of Control Charts Constants - Only intended for developer use}
\usage{
get.norm()
}
\value{
data.frame whose columns are numeric vectors giving constants for control charts limits computation
}
\description{
Creation of Table of Control Chart Constants.
The function is called by \code{\link{control.limits.x}}, \code{\link{control.limits.R}} and \code{\link{control.limits.s}}. This function is not intended for use by regular users of the package.
}
\details{
This function creates a table with columns giving constants for computation limits of control charts.
The function is exported for developer use only. It does not have any input parameters and does not perform any checks on inputs since it is only a convenience function for computing control chart limits.
}
\references{
JOGLEKAR, Anand M. Statistical methods for six sigma: in R&D and manufacturing. Hoboken, NJ: Wiley-Interscience. ISBN sbn0-471-20342-4.
}
