% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{KRDetect.controlchart.plot}
\alias{KRDetect.controlchart.plot}
\title{Control chart outliers detection plot - Only intended for developer use}
\usage{
KRDetect.controlchart.plot(x, all)
}
\arguments{
\item{x}{a list obtained as an output of function \code{\link{KRDetect.outliers.controlchart}} for identification of outliers using control charts.}

\item{all}{a logical variable specifying if individual graphs for outliers detected using control chart x, R and s are plotted together with graph visualising outliers detected based on at least 1 control chart.
If \code{all = FALSE}, only one graph visualising outliers detected based on at least 1 control chart is plotted.}
}
\description{
Plot of results obtained using function \code{\link{KRDetect.outliers.controlchart}} for identification of outliers using control charts.
The function is called by \code{\link{KRDetect.outliers.plot}} and is not intended for use by regular users of the package.
}
\details{
This function plots the results obtained using function \code{\link{KRDetect.outliers.controlchart}} identificating outliers using control charts.
The function is exported for developer use only. It does not perform any checks on inputs since it is only convenience function for plotting results obtained using functions implemented in package \pkg{envoutliers}.
}
