% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{KRDetect.EV.plot}
\alias{KRDetect.EV.plot}
\title{Extreme value outlier detection plot - Only intended for developer use}
\usage{
KRDetect.EV.plot(x, all)
}
\arguments{
\item{x}{a list obtained as an output of function \code{\link{KRDetect.outliers.EV}} for identification of outliers using extreme value theory.}

\item{all}{a logical variable specifying if individual graphs for outliers with extremely low and extremely high value are plotted together with graph visualising outliers with both extremely low and extremely high value.
If \code{all = FALSE}, only one graph visualising outliers with both extremely low and extremely high value is plotted.}
}
\description{
Plot of results obtained using function \code{\link{KRDetect.outliers.EV}} for identification of outliers using extreme value theory.
The function is called by \code{\link{KRDetect.outliers.plot}} and is not intended for use by regular users of the package.
}
\details{
This function plots the results obtained using function \code{\link{KRDetect.outliers.EV}} identificating outliers using changepoint analysis based method.
The function is exported for developer use only. It does not perform any checks on inputs since it is only convenience function for plotting results obtained using functions implemented in package \pkg{envoutliers}.
}
