% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{mc.right}
\alias{mc.right}
\title{Right medcouple (RMC) - Only intended for developer use}
\usage{
mc.right(x)
}
\arguments{
\item{x}{a numeric vector of data values.}
}
\value{
A numeric value giving right medcouple
}
\description{
Calculates right medcouple (RMC).
The function is called by \code{\link{KRDetect.outliers.changepoint}} and is not intended for use by regular users of the package.
}
\details{
This function computes right medcouple (RMC).
The function is exported for developer use only. It does not perform any checks on inputs since it is only a convenience function.
}
\references{
Brys G, Hubert M, Struyf A (2008). Goodness-of-fit tests based on a robust measure of skewness. Computational Statistics, 23(3), 429–442.

Todorov V, Filzmoser P (2009). An Object-Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1-47. URL http://www.jstatsoft.org/v32/i03/.
}
